fix(admission): แก้ type mismatch INT4/INT8 ใน MAX query ที่ทำให้ submit ใบสมัครไม่ได้
CAST(... AS INT) คืนค่า INT4 แต่ Rust กำหนด max_seq: i64 ซึ่ง sqlx คาดหวัง INT8
→ sqlx decode fail → HTTP 500 "Failed to compute application number"

เปลี่ยน CAST(... AS INT) → CAST(... AS BIGINT) และ 0/COALESCE fallback → 0::BIGINT

+++ backend-school/src/modules/admission/handlers/applications.rs
@@ -108,9 +108,9 @@ pub async fn submit_application(
     let max_seq: i64 = sqlx::query_scalar(
         r#"SELECT COALESCE(MAX(
             CASE WHEN application_number ~ '^[0-9]{4}-[0-9]+$'
-            THEN CAST(SPLIT_PART(application_number, '-', 2) AS INT)
-            ELSE 0 END
-        ), 0) FROM admission_applications WHERE application_number LIKE $1"#
+            THEN CAST(SPLIT_PART(application_number, '-', 2) AS BIGINT)
+            ELSE 0::BIGINT END
+        ), 0::BIGINT) FROM admission_applications WHERE application_number LIKE $1"#
     )
     .bind(&year_prefix)
     .fetch_one(&mut *tx)
